try_clean_inventory now only drops items as necessary

+++ computercraft/try_clean_inventory.lua
@@ -10,15 +10,28 @@ function try_clean_inventory()
     local current_slot_index = turtle.getSelectedSlot()
     local empty_slots = 0
 
-    for slot_index=1,16 do
+    for slot_index=1,16 do  -- Count empty slots
         turtle.select(slot_index)
         local item_detail = turtle.getItemDetail()
 
         if item_detail == nil then
             empty_slots = empty_slots + 1
-        elseif constants.non_valuables[item_detail["name"]] then
-            turtle.drop()
-            empty_slots = empty_slots + 1
+        end
+    end
+
+    if empty_slots < constants.empty_slots_required then
+        for slot_index=1,16 do  -- Begin removing junk
+            turtle.select(slot_index)
+            local item_detail = turtle.getItemDetail()
+
+            if item_detail != nil and constants.non_valuables[item_detail["name"]] then
+                turtle.drop()
+                empty_slots = empty_slots + 1
+
+                if empty_slots >= constants.empty_slots_required then  -- Stop once enough slots are empty
+                    break
+                end
+            end
         end
     end
